add admin message protocol,for Example, set the client temperature, byte + byte + int(deviceId) + short(maxTemp) + short(minTemp)

+++ iot-manager-core/src/main/java/com/point/iot/manager/core/facade/MessageDistributorFacade.java
@@ -8,5 +8,5 @@ import java.util.Map;
  * 
  */
 public interface MessageDistributorFacade {
-	Map<Integer, MessageDistributorLogicHandler> getFacadeMap();
+	Map<Byte, MessageDistributorLogicHandler> getFacadeMap();
 }
