+++ iot-manager-core/src/main/java/com/point/iot/manager/core/facade/MessageManagerFacade.java
@@ -7,5 +7,5 @@ import java.util.Map;
  * 
  */
 public interface MessageManagerFacade {
-	Map<Integer, MessageManagerLogicHandler> getFacadeMap();
+	Map<Byte, MessageManagerLogicHandler> getFacadeMap();
 }
